feat : Link Cache Service 타입 구현 및 이름 변경
